+++ frontend/src/constants/route.js
@@ -7,16 +7,7 @@ const ROUTE = {
   },
   MYPAGE: {
     PATH: 'mypage',
-    FOODTRUCKSETTING: { PATH: 'trucksetting/:id' },
-  },
-  FOODLIST: {
-    PATH: 'foodlist',
-    FOODTRUCKS: { PATH: '/' },
-    FOODDETAIL: { PATH: ':id' },
-  },
-  MYPAGE: {
-    PATH: 'mypage',
-    FOODTRUCKSETTING: { PATH: '/' },
+    FOODTRUCKSETTING: { PATH: `trucksetting/:id` },
   },
   LOGIN: { PATH: 'login', NAME: '로그인' },
   LOGOUT: { PATH: 'logout', NAME: '로그아웃' },
