+++ frontend/src/store/atoms.js
@@ -20,16 +20,11 @@ const atoms = {
   }),
 
   menuQuery: atom({ key: 'menu', default: MENU[0].query }),
+
   loginInfo: atom({
     key: 'loginInfo',
     default: { storeId: null, localId: null },
     effects: [sessionStorageEffect('storeId')],
-  }),
-
-  isLogin: atom({
-    key: 'isLogin',
-    default: { state: false, type: null },
-    effects: [sessionStorageEffect('isLogin')],
   }),
 
   modal: atom({
